Updated the installer and the application.
The installer now has it's own logo (Re-Co logo) and can also create a desktop shortcut for us. There was a tiny issue with ths application where bringing it back up after closing it using the system tray icon will result in not showing the windows properly. This issue is fixed, although not stable.